feat: Create action in nav graph
