--- app/src/main/java/com/example/ynov_lyon_bde/PersonalInformationsUser.kt
@@ -1,17 +1,0 @@
-package com.example.ynov_lyon_bde
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-class PersonalInformationsUser : Fragment() {
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
-        // Inflate the layout for this fragment
-        val view = inflater.inflate(R.layout.fragment_personal_informations_user, container, false)
-        return view
-    }
-}
