+++ app/src/main/res/layout-hdpi/fragment_settings_user_profile.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/ic_sign_in_background"
-    tools:context=".ui.screens.profil.SettingsUserProfil">
+    tools:context=".ui.screens.profil.SettingsUserProfile">
 
     <!-- Title settings bloc -->
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -25,8 +25,8 @@
             android:layout_height="match_parent"
             android:paddingStart="16dp"
             android:scaleType="fitXY"
-            android:scaleX="1.3"
-            android:scaleY="1.3"
+            android:scaleX="1.1"
+            android:scaleY="1.1"
             android:src="?attr/actionModeCloseDrawable"
             app:tint="#000000"
             app:layout_constraintBottom_toBottomOf="parent"
@@ -44,7 +44,7 @@
             android:padding="16dp"
             android:text="@string/fragment_settings_user_profil_title"
             android:textColor="#000"
-            android:textSize="24sp"
+            android:textSize="20sp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toEndOf="@+id/back_icon"
@@ -56,7 +56,6 @@
         android:id="@+id/settings_fields_bloc"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="32dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/title_bloc"
         app:layout_constraintStart_toStartOf="parent">
@@ -65,7 +64,7 @@
         <com.google.android.material.button.MaterialButton
             android:id="@+id/personalInformationButton"
             android:layout_width="match_parent"
-            android:layout_height="66dp"
+            android:layout_height="46dp"
             android:layout_marginTop="26dp"
             android:layout_marginStart="20dp"
             android:layout_marginEnd="20dp"
@@ -76,7 +75,7 @@
             android:hint="@string/personal_information"
             android:drawableTint="#707070"
             android:textAlignment="textStart"
-            android:textSize="18sp"
+            android:textSize="12sp"
             android:fontFamily="@font/roboto_regular"
             app:layout_constraintBottom_toTopOf="@+id/howUseAppButton"
             app:layout_constraintEnd_toEndOf="parent"
@@ -87,7 +86,7 @@
         <com.google.android.material.button.MaterialButton
             android:id="@+id/howUseAppButton"
             android:layout_width="match_parent"
-            android:layout_height="66dp"
+            android:layout_height="46dp"
             android:layout_marginStart="20dp"
             android:layout_marginTop="26dp"
             android:layout_marginEnd="20dp"
@@ -98,7 +97,7 @@
             android:hint="@string/how_use_app"
             android:drawableTint="#707070"
             android:textAlignment="textStart"
-            android:textSize="18sp"
+            android:textSize="12sp"
             android:fontFamily="@font/roboto_regular"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
@@ -108,7 +107,7 @@
         <com.google.android.material.button.MaterialButton
             android:id="@+id/payementInformationButton"
             android:layout_width="match_parent"
-            android:layout_height="66dp"
+            android:layout_height="46dp"
             android:layout_marginStart="20dp"
             android:layout_marginTop="26dp"
             android:layout_marginEnd="20dp"
@@ -119,7 +118,7 @@
             android:hint="@string/payment_information"
             android:drawableTint="#707070"
             android:textAlignment="textStart"
-            android:textSize="18sp"
+            android:textSize="12sp"
             android:fontFamily="@font/roboto_regular"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
@@ -129,7 +128,7 @@
         <com.google.android.material.button.MaterialButton
             android:id="@+id/deleteAccountButton"
             android:layout_width="match_parent"
-            android:layout_height="66dp"
+            android:layout_height="46dp"
             android:layout_marginStart="20dp"
             android:layout_marginTop="26dp"
             android:layout_marginEnd="20dp"
@@ -140,7 +139,7 @@
             android:hint="@string/delete_account"
             android:drawableTint="#707070"
             android:textAlignment="textStart"
-            android:textSize="18sp"
+            android:textSize="12sp"
             android:fontFamily="@font/roboto_regular"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
@@ -150,7 +149,7 @@
         <com.google.android.material.button.MaterialButton
             android:id="@+id/logoutButton"
             android:layout_width="match_parent"
-            android:layout_height="66dp"
+            android:layout_height="46dp"
             android:layout_marginStart="20dp"
             android:layout_marginTop="26dp"
             android:layout_marginEnd="20dp"
@@ -160,7 +159,7 @@
             android:fontFamily="@font/roboto_regular"
             android:hint="@string/logout"
             android:drawableTint="#707070"
-            android:textSize="18sp"
+            android:textSize="12sp"
             android:textAlignment="textStart"
             app:backgroundTint="@color/settings_choice"
             app:layout_constraintBottom_toBottomOf="parent"
