+++ app/src/main/res/layout/fragment_personal_informations_user.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PersonalInformationsUser">
+    tools:context=".ui.screens.profil.PersonalInformationsUser">
 
 </FrameLayout>
